[MOREL-46] Core language
Add an intermediate language between AST and the interpreted
language (Code/Applicable). The language mirrors the AST
language, but is simpler, and every expression/pattern node
knows its type.

Examples:
 * There is no class Core.If; Ast.If becomes Core.Case;
 * There is no class Core.ListExp; list expressions become
   calls to a list-constructor function;
 * There is no class Core.ApplicableExp; constant functions
   now become function literals (instances of Core.Literal
   with op = FN_LITERAL).

Function literals now use BuiltIn, not Op.

There are also value literals (instances of Core.Literal
with op = VALUE_LITERAL) for non-primitive values.

Core.RecordSelector contains the ordinal of the field, not
its name. This possible because Core expressions are typed.

Rename class Ast.LetExp to Let (and similarly Core.Let).

Add class Inliner, a visitor that recognizes and expands
constant expressions. This applies not just to primitive
expressions (int, bool, string etc.) but also to structured
expressions (lambdas, records, lists). We currently stop
short of closures (because are printed as very verbose
plans) but hope to revisit.

Inlining is a key technique in optimizing functional
programs.

Inlining is sometimes reversible. Consider the function
'String.extract'. It is represented as '#extract String',
that is, the 'extract' field of the built-in 'String'
structure, which is a record most of whose fields are
functions. During inlining it becomes a function literal
that references BuiltIn.STRING_EXTRACT. But when we want to
convert Core code back to a Morel string, we can do a
reverse lookup, and generate '#extract String'.