Analyze uses of bindings
How bindings are used informs whether they can be inlined.
For example, if a binding 'val x = y + z' is used twice
then inlining it will cause more work and larger code. If
'x' is used in separate branches, such as 'if a then x else
x * 2', it is the same work but larger code. There are
other cases where inlining may increase work but does not
increase code size.

The paper 'Secrets of the Glasgow Haskell Compiler
inliner' (Peyton Jones, Marlow, 2002) inspired the rules
for when it is safe to inline.