[MOREL-179] In compilation environment, use name + ordinal as the key, not just name, to accommodate variable copies caused by inlining
Add class RefChecker to detect bad references in FromBuilder.

In EnvVisitor, improve the environment passed to Group;
in EnvVisitor and EnvShuttle, RecValDecl adds to the
environment in which the expression body is processed.

Generate fewer new variables when copying a Yield step.

When copying a Yield step like '{a = exp1, b = exp2}', inherit
bindings from the previous version of the step so that the
'a' and 'b' varibles have the same ordinal, and will match
the variables needed by the next step.

We can remove a trivial Yield, but 'yield {e=e}' isn't
trivial if the output IdPat has a different ordinal than the
input IdPat.

Fixes #179