Add "directory" property to Session
The directory is used by the "use" command (for running
scripts) and the "file" value (for opening files as tables).

When launched via the "morel" script the default directory is
the current directory, but can be overridden using the
"--directory" flag.

Allow properties (class Prop) to be of type String or File
(in addition to Integer and Boolean).