Start [MOREL-xx] Add a bag type, and Bag structure, to represent unordered collections with duplicates allowed
The built-in `list` type represents an ordered collection with duplicates allowed; in this change, we add a `bag` type that also allows duplicates but has no guaranteed order.

This means that, given a bag and a list with the same N elements, the list contains more information, because it has remembered which of the N! orderings of the elements has occurred. Converting from a list to a bag is therefore lossy. But there is a bijection between 'a list and ('a * int) bag provided that the int ordinals are unique.

The Bag structure has analogous operations to the standard List structure, except for nth, take, drop, rev.

We add mapi to List (based on Vector.mapi).

Fixes #235