Wip 12/27 (13 failures, 19 errors)
From expression used to be ListType, now it must be a collection
type (ListType or a bag). Add Type.isCollection().

Change built-in relational functions (e.g. Relational.only)
to accept bag rather than list argument. (In future we will
automatically convert list to bag when calling aggregate
functions, and perhaps when using a 'from' expression as
argument to a function that requires a list.)