[MOREL-245] Enable google-java-format
`google-java-format` reformats code to a standard style.
It reduces merge conflicts because all PRs have the same
style.

Since google-java-format is reformatting the code, less is
needed of checkstyle, but we still need it for semantic
checks, such as that files have license headers.

`LintTest` (added in #200) still performs its checks.

Fixes #245