[MOREL-252] When unparsing Morel, try to generate qualified identifiers less often
An identifier (`class Core.Id`) has not just a name but a
version, so that we can distinguish identifiers that have
the same name but different identities. (The previous
version is obscured in the environment, but may still be
referenced via closures.)

Suppose an identifier `x` always occurs with ordinal 2 in
this expression but occurs with ordinals 0, 1 and 3
elsewhere in the program. Previously we would output `x_2`,
but after this change, `AstWriter` remembers (in a map)
that 2 is the only version of `x` that we have seen, and
output it as just `x`; if it later sees `x` with ordinal 3,
it will output `x_1`.

Fixes #252