Start [MOREL-262] Conversion of distinct does not preserve types
This commit tries to change the rules about whether a scalar yield is
allowed if it's not last.

And with these changes in the rules, the query

  from i in [1, 2, 3, 4, 5, 6],
      j in [1, 2, 3, 4, 5, 6]
    yield i + j
    distinct

would be regarded as invalid.

But I think that's wrong. We should let people write 'yield z = i + j',
then refer to 'z', e.g. in 'yield z = i + j order by z'. And generate a
sensible implicitYieldExpression for queries like '... yield i + j
distinct'.

Previous work is #213 and commit 33ae519.