Refactor `TypeResolver`
Import the inner classes of `Unifier`, which are used
throughout `TypeResolver`.

Rename `class Foo` to `KeyBuilder`.

Re-order args of `equiv()`; add overloads of `reg()` with
and without variable.

Convert `termMap` and `fieldVars` from `Map` to `PairList`.

Add methods to create terms, such as `listTerm()`.

Factor out methods `deduceQueryType`, `deduceYieldType`,
`deduceScanStepType`, and `deduceGroupStepType`.

In `MainTest`, some tests in `testFromYield` and
`testYieldSingletonRecord` are commented out, and there are
failures in `relational.smli` and `suchThat.smli`.