[MOREL-276] Add `ordinal`
Following #273, we add an `ordinal` expression to capture
the current position in an ordered iteration as a data
value. It is an integer starting from 0.

In the following query, the query becomes unordered after
the join with the unordered `scott.depts` table, but we can
restore the same order by sorting on `i` (which is assigned
in a previous step from `ordinal`).
```
from e in scott.emps
  order e.deptno
  yield {i = ordinal, e}
  join d in scott.depts on d.deptno = e.deptno
  order i
  yield {i, d.dname, e.ename};
> val it =
>   [{dname="ACCOUNTING",ename="CLARK",i=1},
>    {dname="ACCOUNTING",ename="KING",i=2},
>    {dname="ACCOUNTING",ename="MILLER",i=3},
>    {dname="RESEARCH",ename="SMITH",i=4},
>    {dname="RESEARCH",ename="JONES",i=5},
>    {dname="RESEARCH",ename="SCOTT",i=6},
>    {dname="RESEARCH",ename="ADAMS",i=7},
>    {dname="RESEARCH",ename="FORD",i=8},
>    {dname="SALES",ename="ALLEN",i=9},
>    {dname="SALES",ename="WARD",i=10},
>    {dname="SALES",ename="MARTIN",i=11},
>    {dname="SALES",ename="BLAKE",i=12},
>    {dname="SALES",ename="TURNER",i=13},
>    {dname="SALES",ename="JAMES",i=14}]
>   : {dname:string, ename:string, i:int} list
```

Morel's order step does not provide stable sorts, but you
can achieve stable sorts using `ordinal`.
```
from e in scott.emps
  order e.empno
  where e.deptno > 10
  order e.job, ordinal;
```
When ordering by job there are ties, so adding ordinal as
the tie-breaker ensures that if two employees have the
same job, whichever came earlier in the input will come
earlier in the output.