Add syntax to test scripts for type-checking commands
If you write
  1;
  > val it = 1 : int
it is executed, but if you prefix with ':t', like this:
  :t 1;
  > val it : int
it only its type is derived.

Note that ':t' can occur on its own line, which is useful if the
statement is intended to produce warnings with positions. And it
can occur in the middle of a statement. (You should only do this
if the statement is preceded by a comment.)