Invoke `.smli` scripts from the command line
You can invoke via Maven like this:

  ./mvnw -q test-compile exec:java \
      -Dexec.mainClass="net.hydromatic.morel.Script" \
      -Dexec.classpathScope=test \
      -Dexec.args="--echo script/wordle.smli

Factor a new `class Script` out of `ScriptTest`. For each
`.smli` file, `ScriptTest` generates a test, and the test
runs by instantating and invoking a `Script`. Move the
`ScriptTest.main()` method to `Script`, so this is now the
way to invoke a script from the command line.

Add an `--echo` flag to `Script.main`. The `--echo` option
causes output to be printed to both the log file and stdout,
and makes it easier to find sections of a script that are
slow or hang.