Datalog (#323)
Add a Datalog interface to Morel.

A new package `net.hydromatic.morel.datalog` contains a Datalog
parser written in JavaCC. Here is a sample program:

  .decl parent(ancestor: string, child: string)
  .decl ancestor(ancestor: string, descendant: string)

  parent("earendil", "elrond").
  parent("elrond", "arwen").
  parent("elrond", "elladan").
  parent("elrond", "elrohir").

  ancestor(X, Y) :- parent(X, Y).
  ancestor(X, Z) :- ancestor(X, Y), ancestor(Y, Z).

  .output ancestor

The syntax is based on Souffle, except that Souffle's `symbol`
and `number` types are changed to `string` and `int` to match
Morel's types, and the `.input` directive has an extra optional
`path` argument. Arithmetic in rule heads and bodies is supported;
aggregate functions, `contains`, `match`, `true`, and `false` are
not yet supported. Another possible enhancement would be to allow
any Morel table (list or bag of records) to be used as a Datalog
relation.

The Datalog layer gives compile-time errors for programs that are
unsafe or not stratified.

The functionality is packaged in a new `Datalog` structure, with
methods including `execute: string -> variant`, `validate`, and
`translate`.

A new test script `datalog.smli` calls `Datalog.execute` with
various programs, including each of the Datalog examples in
`such-that.smli`.

Extend `CLAUDE.md` with general details of Morel and its
development process, useful for humans and machines alike.

Fixes #323